Move ensure trace-correlation test into index.test.ts
